Mensageria integrada sqs e sns com assinatura

+++ src/main/java/com/borba/storage/sqs/SqsConsumer.java
@@ -1,6 +1,8 @@
 package com.borba.storage.sqs;
 
 import com.borba.storage.configuration.AwsProperties;
+import com.borba.storage.service.PresignedUrlService;
+import com.borba.storage.sns.SnsPublisher;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import software.amazon.awssdk.services.sqs.SqsClient;
@@ -8,23 +10,29 @@ import software.amazon.awssdk.services.sqs.model.DeleteMessageRequest;
 import software.amazon.awssdk.services.sqs.model.Message;
 import software.amazon.awssdk.services.sqs.model.ReceiveMessageRequest;
 
+import java.time.Duration;
 import java.util.List;
 
 @Service
 public class SqsConsumer {
 
-    private SqsClient sqsClient;
-    private String queueUrl;
+    private final  SqsClient sqsClient;
+    private final String queueUrl;
+    private final PresignedUrlService presignedUrlService;
+    private final SnsPublisher snsPublisher;
 
-    public SqsConsumer(SqsClient sqsClient, AwsProperties props) {
+    public SqsConsumer(SqsClient sqsClient, AwsProperties props,PresignedUrlService presignedUrlService,
+                       SnsPublisher snsPublisher) {
         this.sqsClient = sqsClient;
         this.queueUrl = props.getSqs().getQueueUrl();
+        this.presignedUrlService = presignedUrlService;
+        this.snsPublisher = snsPublisher;
     }
 
     @Scheduled(fixedDelay = 5000)
     public void pullMessages(){
 
-        System.out.println("Verificando fila...");
+        //System.out.println("Verificando fila...");
 
         List<Message> messages = sqsClient.receiveMessage(
                 ReceiveMessageRequest.builder()
@@ -37,8 +45,26 @@ public class SqsConsumer {
         for (Message msg : messages){
 
             try {
+
+
                 //Processar mensagem
+
+                String body = msg.body();
+
                 System.out.println("Recebida: " + msg.body());
+
+                // Extrair a chave do arquivo da mensagem
+                String key = extractKey(body);
+                if (key == null || key.isBlank()) {
+                    System.err.println("Mensagem inválida: chave não encontrada.");
+                    continue;
+                }
+
+                // Gerar URL válida por 15 minutos
+                String url = presignedUrlService.gerarPresignedUrl(key, Duration.ofMinutes(15));
+
+                String mensagem = "Seu arquivo está pronto para download: " + url;
+                snsPublisher.publishMessage(mensagem);
 
                 sqsClient.deleteMessage(DeleteMessageRequest.builder()
                         .queueUrl(queueUrl)
@@ -50,5 +76,13 @@ public class SqsConsumer {
             }
 
         }
+    }
+
+    private String extractKey(String body) {
+        // Exemplo simples: "Arquivo enviado com sucesso. Key: nome-do-arquivo.txt"
+        if (body.contains("Key:")) {
+            return body.substring(body.indexOf("Key:") + 5).trim();
+        }
+        return null;
     }
 }
